refactor: move theme dir to the store/themes
rename  directory: theme -> themes
